+++ wass2s/was_machine_learning.py
@@ -2479,12 +2479,6 @@ class WAS_MARS_Model:
             Upper tercile threshold.
         dist_code : int
             Distribution code (1-8).
-        shape : float
-            Shape parameter.
-        loc : float
-            Location parameter.
-        scale : float
-            Scale parameter.
     
         Returns
         -------